<Feat> : change fixture dir
